third_party/blink/web_tests/fast/peerconnection/RTCPeerConnection-garbagecollected.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../resources/gc.js"></script>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
</head>
<body>
<script>

const limit = 10;

promise_test(async t => {
  const baseline = internals.peerConnectionCount();
  let pc;
  for (let count = 0; count < limit; count++) {
    pc = new RTCPeerConnection();
    pc.close();
  }
  await asyncGC();
  // All the non-referenced PCs should now be collected.
  assert_equals(internals.peerConnectionCount(), baseline + 1, "Failed at Check 1, baseline was " + baseline);
  pc = null;
  await asyncGC();
  assert_equals(internals.peerConnectionCount(), baseline, "Failed at Check 2, baseline was " + baseline);
}, 'Closed Peerconnections are garbage collected');


promise_test(async t => {
  const baseline = internals.peerConnectionCount();
  for (let count = 0; count < limit; count++) {
    pc = new RTCPeerConnection();
  }
  pc = null;
  await asyncGC();
  assert_equals(internals.peerConnectionCount(), baseline + limit, "Failed - Baseline was " + baseline);
}, 'Non-closed Peerconnections are NOT garbage collected');

</script>
</body>
</html>
